Backend_manual file was added. Endpoints were described. Some little changes were made

+++ application/dtos/backgroundWorkerSettingsDto.go
@@ -3,7 +3,7 @@ package dtos
 type BackgroundWorkerSettingsDto struct {
 	WorkerName         string  `json:"workerName"`
 	Timer              float64 `json:"timer"`
-	RequestDelay       int     `json:"requestDellay"`
+	RequestDelay       float64 `json:"requestDellay"`
 	Random             bool    `json:"random"`
 	WritesNumberToSend int     `json:"writesNumberToSend"`
 	TotalToSend        int64   `json:"totalToSend"`
