+++ core/models/backgroundworker.go
@@ -7,7 +7,7 @@ type Work func(ctx context.Context) int
 type BackgroundWorkerSettings struct {
 	WorkerName         string           `json:"workerName"`
 	Timer              float64          `json:"timer"`
-	RequestDelay       int              `json:"requestDellay"`
+	RequestDelay       float64          `json:"requestDellay"`
 	Random             bool             `json:"random"`
 	WritesNumberToSend int              `json:"writesNumberToSend"`
 	TotalToSend        int64            `json:"totalToSend"`
